Safari-609.1.11.1/LayoutTests/http/tests/security/contentSecurityPolicy/source-list-parsing-paths-02-expected.txt

CONSOLE MESSAGE: The source list for Content Security Policy directive 'script-src' contains a source with an invalid path: '/security/#query=string'. The fragment identifier, including the '#', will be ignored.
CONSOLE MESSAGE: The source list for Content Security Policy directive 'script-src' contains a source with an invalid path: '/security/?query=string'. The query component, including the '?', will be ignored.
CONSOLE MESSAGE: The source list for Content Security Policy directive 'script-src' contains a source with an invalid path: '/not-security/#query=string'. The fragment identifier, including the '#', will be ignored.
CONSOLE MESSAGE: Refused to load http://127.0.0.1:8000/security/contentSecurityPolicy/resources/script.js because it does not appear in the script-src directive of the Content Security Policy.
CONSOLE MESSAGE: The source list for Content Security Policy directive 'script-src' contains a source with an invalid path: '/not-security/?query=string'. The query component, including the '?', will be ignored.
CONSOLE MESSAGE: Refused to load http://127.0.0.1:8000/security/contentSecurityPolicy/resources/script.js because it does not appear in the script-src directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to load http://127.0.0.1:8000/security/contentSecurityPolicy/resources/script.js because it does not appear in the script-src directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to load http://127.0.0.1:8000/security/contentSecurityPolicy/resources/script.js because it does not appear in the script-src directive of the Content Security Policy.
Resources should be rejected unless they match a whitelisted path.



--------
Frame: '<!--frame1-->'
--------
PASS

--------
Frame: '<!--frame2-->'
--------
PASS

--------
Frame: '<!--frame3-->'
--------
PASS

--------
Frame: '<!--frame4-->'
--------
PASS

--------
Frame: '<!--frame5-->'
--------
PASS

--------
Frame: '<!--frame6-->'
--------
PASS

--------
Frame: '<!--frame7-->'
--------
PASS

--------
Frame: '<!--frame8-->'
--------
PASS
